% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_events}
\alias{drug_events}
\title{This is an example dataset of the FDA adverse drug events.
It contains the FDA adverse drug event data from Q1 of 2018
For a more specific dataset, use the function get_tables()}
\format{An object of class \code{list} of length 2.}
\source{
\url{http://open.fda.gov}
}
\usage{
drug_events
}
\description{
This is an example dataset of the FDA adverse drug events.
It contains the FDA adverse drug event data from Q1 of 2018
For a more specific dataset, use the function get_tables()
}
\examples{
data(drug_events)
}
\keyword{datasets}
