% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda.drug.stats290.R
\docType{package}
\name{fda.drug.stats290}
\alias{fda.drug.stats290}
\alias{fda.drug.stats290-package}
\title{Stat 290 FDA package}
\description{
This package provides an easy wrapper over the Open FDA API. 
The package simplifies calls to obtain data, provides ways to reshape data,
and helps users count values in fields. The aim of this package is to provide
a simpler way for researchers to explore and understand the dataset.
}
