% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_api_table.R
\name{fda_api_table}
\alias{fda_api_table}
\title{Load JSON tables from the adverse drug event}
\usage{
fda_api_table(limit = 1, api_key = NULL, start_date = NULL,
  end_date = NULL, reactionmeddrapt = NULL, actiondrug = NULL,
  patientsex = NULL, patientagegroup = NULL, reactionoutcome = NULL,
  qualification = NULL, generic_name = NULL, brand_name = NULL,
  serious = NULL, congenitalanomali = NULL, death = NULL,
  disabling = NULL, hospitalization = NULL, lifethreatening = NULL)
}
\arguments{
\item{limit}{number of data to pull - if it is not set, return only one instance
limit should be less than 100}

\item{api_key}{include an API key for faster querying. See the documentation 
for more details.}

\item{start_date}{the start date of events you want to search}

\item{end_date}{the end date of the events you want to search}

\item{reactionmeddrapt}{the patient reaction
(encoded in British English, e.g. diarrhea spelled as diarrohea)}

\item{actiondrug}{actions taken with the drug:
1 = drug withdrawn, 2 = dose reduced, 3 = dose increased, 4 = dose unchanged
5 = unknown 6 = not applicable}

\item{patientsex}{the sex of the patient, where 0 = unknown, 1 = male,
2 = female}

\item{patientagegroup}{patient age group code - 1 = neonate, 2 = infant,
3 = child, 4 = adolescent, 5 = adult, 6 = elderly, default set to NULL (use for filtering results)}

\item{reactionoutcome}{outcome of reaction at the time of last
observation. 1 = recovered, 2 = recovering, 3 = not recovered, 4 = recovered with
health issues, 5 = fatal, 6 = unknown}

\item{qualification}{category of individual who submitted report
1 = physician, 2 = pharmacist, 3 = other health professional, 4 = lawyer, 5 = consumer}

\item{generic_name}{generic name of product}

\item{brand_name}{brand/trade name of drug}

\item{serious}{whether the drug event was serious 
(1 = life threatening illness, hospitalization, fatality, etc., 2 = otherwise)}

\item{congenitalanomali}{1 if resulted in congenital anomali, otherwise N/A}

\item{death}{1 if resulted in death, otherwise N/A}

\item{disabling}{1 if resulted disabilities, otherwise N/A}

\item{hospitalization}{1 if resulted in hospitalization, otherwise N/A}

\item{lifethreatening}{1 if resulted in life threatening conditions, otherwise, N/A}
}
\value{
a dataset of adverse drug events
}
\description{
Load data from the FDA adverse drug events table
}
\examples{
t <- fda_api_table()
fda_api_table(
             limit = 100, 
             start_date = "2016-01-01", 
             end_date = "2018-01-01",
             reactionmeddrapt = "nephrolithiasis",
             actiondrug = 5,
             patientsex = 1,
             reactionoutcome = 6,
             qualification = 1,
             generic_name  = "aspirin"
             )
}
\keyword{FDA}
