% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_dataframe.R
\name{fda_dataframe}
\alias{fda_dataframe}
\title{Transform JSON tables exported by function fda_api_table to dataframe (not including drug information; 
for accessing to drug information in data frame, see fda_drug_dataframe.}
\usage{
fda_dataframe(g_table)
}
\arguments{
\item{g_table}{a JSON table returned by fda_api_table}
}
\value{
a data frame with all variables from FDA adverse drug events except drug information.
}
\description{
Transform JSON tables exported by the function fda_api_table into data frame; not including drug information; 
for accessing to drug information in data frame, see fda_drug_dataframe.
}
\examples{
g_table <- fda_api_table()
fda_dataframe(g_table)
}
\keyword{FDA}
