% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_drug_dataframe.R
\name{fda_drug_dataframe}
\alias{fda_drug_dataframe}
\title{Extract drug information from JSON tables exported by function fda_api_table
to data frame not including openfda information; for accessing to openfda 
information, see get_openfda.}
\usage{
fda_drug_dataframe(g_table, id = NULL)
}
\arguments{
\item{g_table}{a JSON table returned by fda_api_table}

\item{id}{an integer vector, for subsetting patient observations by specifying 
their IDs. If not specified, return drug information of all patients passed by 
g_table.}
}
\value{
a data frame with all drug information of specified patients from FDA 
adverse drug events except openfda.
}
\description{
Extract drug information from JSON tables exported by function
fda_api_table to data frame not including openfda information; for accessing 
to openfda information, see get_openfda.
}
\examples{
g_table <- fda_api_table(limit=100, patientsex = 1)
fda_drug_dataframe(g_table)
}
\keyword{FDA}
