% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfda_count.R
\name{openfda_count}
\alias{openfda_count}
\title{Return a data frame of unique elements of a specified variable in a list of openfda
data and their frequencies. The data frame is sorted according to frequencies of each element in
descending order.}
\usage{
openfda_count(var_name, openfda_data)
}
\arguments{
\item{var_name}{a string for the column/variable name}

\item{openfda_data}{an openfda_data list returned by get_openfda}
}
\value{
a dataframe of element names and their frequencies sorted according 
to frequencies of each element in descending order
}
\description{
Return a data frame of unique elements of a specified variable in a list of openfda
data and their frequencies. The data frame is sorted according to frequencies of each element in
descending order.
}
\examples{
g_table <- fda_api_table(limit=100, patientsex = 1)
openfda_data <- get_openfda(g_table)
openfda_count("route", openfda_data)
}
\keyword{FDA}
