% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_top_react.R
\name{view_top_react}
\alias{view_top_react}
\title{Return a data frame with activesubstance and its corresponding patient reaction, and
show histograms of the counts of reactions of each activesubstance.}
\usage{
view_top_react(plain_fda, drug_fda, top_name = 5, top_react = 8)
}
\arguments{
\item{plain_fda}{a data frame returned by fda_dataframe}

\item{drug_fda}{a data frame returned by fda_drug_dataframe}

\item{top_name}{integer, specifying the number of activesubstance names for the output; this number of
activesubstance names with highest frequencies will be taken; default = 5}

\item{top_react}{integer, specifying the number of patient reaction types for the output; this number of
patient reaction types with highest frequencies will be taken; default = 8}
}
\value{
a data frame with activesubstance and its corresponding patient reaction
}
\description{
Return a data frame of two columns: one is activesubstance from drug_fda 
data frame, with top frequencies specified by top_name; the other is the corresponding 
patient reaction from plain_fda data frame, with top frequencies specified by top_react.
At the same time, it prints out the bar plots with ggplot2 library of each activesubstance
and its counts of patient reactions.
}
\examples{
try_dt <- fda_api_table(limit=100)
fda_data<-fda_dataframe(try_dt)
fda_drug <- fda_drug_dataframe(try_dt)
view_top_react(fda_data, fda_drug, top_name = 5, top_react = 8)
}
\keyword{FDA}
